backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
werkzeug==2.3.7